move to decide with other torrent_cache functions
