+++ src/cmd.ts
@@ -3,7 +3,7 @@ import chalk from "chalk";
 import { Option, program } from "commander";
 import { inspect } from "util";
 import { getApiKey, resetApiKey } from "./auth.js";
-import { updateTorrentCache } from "./cache.js";
+import { updateTorrentCache } from "./decide.js";
 import { customizeErrorMessage, VALIDATION_SCHEMA } from "./configSchema.js";
 import { FileConfig, generateConfig, getFileConfig } from "./configuration.js";
 import {
@@ -363,19 +363,19 @@ program
 	});
 
 program
-	.command("update-torrent-cache-urls")
+	.command("update-torrent-cache-trackers")
 	.description("Update announce urls in the torrent cache")
 	.usage("<old-announce-url> <new-announce-url>")
 	.argument(
 		"old-announce-url",
-		'A substring of the announce url to replace, e.g. update-torrent-cache-urls "old.example.com" "new.example.com"',
+		'A substring of the announce url to replace, e.g. update-torrent-cache-trackers "old.example.com" "new.example.com"',
 	)
 	.argument(
 		"new-announce-url",
-		'A substring of the new announce url to replace the old one with, e.g. update-torrent-cache-urls "myoldpasskey" "mynewpasskey"',
+		'A substring of the new announce url to replace the old one with, e.g. update-torrent-cache-trackers "myoldpasskey" "mynewpasskey"',
 	)
-	.action(async (oldAnnounceUrl, newAnnounceUrl) => {
-		await updateTorrentCache(oldAnnounceUrl, newAnnounceUrl);
+	.action(async (oldStr, newStr) => {
+		await updateTorrentCache(oldStr, newStr);
 		await db.destroy();
 		await memDB.destroy();
 	});
